+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
+
+
+### Changed
+* Update tests for dev-site v6
 
 5.25.0 - (July 30, 2019)
 ------------------