+++ packages/terra-form-select/CHANGELOG.md
@@ -7,6 +7,19 @@ Unreleased
 * Added aria-label to resolve accessibility violation with axe-core v3.3.0.
 * Updated translations.
 
+### Changed
+* updated package.json test scripts
+
+
+### Changed
+* Update tests for dev-site v6
+
+5.25.0 - (July 30, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+
 5.24.0 - (July 24, 2019)
 ------------------
 ### Changed